% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{cfa_from_matrix}
\alias{cfa_from_matrix}
\title{Confirmatory Factor Analysis}
\usage{
cfa_from_matrix(X, selectedContinuousID, selectedBinaryID, adjM)
}
\arguments{
\item{X}{A matrix (n-by-d) containing n data samples of d variables}

\item{selectedContinuousID}{Vector of indices of continuous variables}

\item{selectedBinaryID}{Vector of indices of binary or ordinal categorical variables}

\item{adjM}{The adjacency matrix between indicators and latent variables, value of 1 indicate a connection}
}
\value{
A list consisting of correlation matrix, loading factors  and fit measures
}
\description{
Performs confirmatory factor analysis on a user-specified network
}
