% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{cfa_from_matrix}
\alias{cfa_from_matrix}
\title{Confirmatory Factor Analysis}
\usage{
cfa_from_matrix(X, selectedContinuousID, selectedBinaryID, adjM,
  lav_estimator = "default", predict_method = "EBM", verbose = FALSE)
}
\arguments{
\item{X}{A matrix (n-by-d) containing n data samples of d variables}

\item{selectedContinuousID}{Vector of indices of continuous variables}

\item{selectedBinaryID}{Vector of indices of binary or ordinal categorical variables}

\item{adjM}{The adjacency matrix between indicators and latent variables, value of 1 indicate a connection}

\item{lav_estimator}{The estimator to be used. Can be one of the following: "ML" for maximum likelihood, "GLS" for generalized leastsquares, "WLS" for weighted least squares (sometimes called ADF estimation), "ULS" for unweighted least squares and "DWLS" for diagonally weighted least squares. These are the main options that affect the estimation. For convenience, the "ML" option can be extended as "MLM", "MLMV", "MLMVS", "MLF", and "MLR"}

\item{predict_method}{A character string. In the linear case (when the indicators are continuous), the possible options are "regression" or "Bartlett". In the categorical case, the two options are "EBM" for the Empirical Bayes Modal approach, and "ML" for the maximum likelihood approach.}

\item{verbose}{TRUE or FALSE, whether to output debug information.}
}
\value{
A list consisting of correlation matrix, loading factors, factor scores  and fit measures
}
\description{
Performs confirmatory factor analysis on a user-specified network
}
